added  some ODs for L3(8) (and extensions) mod 2
The characters of degrees 8 and 64 had been missing from the list
because their indicators in the character tables were
erroneously stated as "o".
Now the character tables have been corrected,
and the list has been extended accordingly.
Since the representations in question are available in the AGR,
the OD values are now known.