fixed an inconsistency with the 990a of A11
Consider A_{12}, p = 0, 1650a (a rational character).

On the one hand, the restriction to A_{11} decomposes as 660a + 990a,
with stored ODs 5 * 21 = 105.

On the other hand, the restriction to (A_9 \times 3):2 decomposes as
5 + 13 + 22 + 23 + 25 + 43 + 44 + 45 + 46,
with ODs
-15 * 105 * 21 * -7 * 1 * -3 * 1 * 1 * 1 = -7 (mod squares).
(Some of these values are stored for S9, the others are computed
via the eigenvalues criterion.)

I think that the stored OD of the 990a of A_{11} is not correct.
If we restrict the 990a to (A_8 \times 3):2 then we get
7 + 11 + 12 + 17 + 19 + 20 + 21 + 26 + 27 + 30 + 34 + 36,
with ODs
21 * -15 * 5 * 5 * 105 * 105 * -3 * 1 * -3 * 1 * 1 * 1 = -35 (mod squares).

With this value, we get OD -7 for 1650a of A_{12}.

Note that Richard had claimed OD -7 for 1650a of A_{12} but on March 30th
I had proposed to change the value,
based on the (now apparently wrong) value 21 for the 990a of A_{11}.